% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_uni.R
\name{.explore_uni_quali}
\alias{.explore_uni_quali}
\title{Qualitative variable representation}
\usage{
.explore_uni_quali(x)
}
\arguments{
\item{x}{:\code{character/vector}, variable to show}
}
\value{
a \code{ggplot2} object
}
\description{
Qualitative variable representation
}
