% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_uni.R
\name{.explore_uni_quanti}
\alias{.explore_uni_quanti}
\title{Quantitative variable representation}
\usage{
.explore_uni_quanti(x)
}
\arguments{
\item{x}{: \code{numeric/vector}, variable to show}
}
\value{
a \code{ggplot2} object
}
\description{
Quantitative variable representation
}
