% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{exploredata.env}
\alias{exploredata.env}
\title{Define 1 export a new custom environnement...}
\format{
An object of class \code{environment} of length 0.
}
\usage{
exploredata.env
}
\description{
Define 1 export a new custom environnement...
}
\keyword{datasets}
