% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exploredata_ex}
\alias{exploredata_ex}
\title{Demo data for exploredata}
\format{
A data frame with :
\describe{
\item{value}{quantitative variable}
\item{group}{qualitative variable}
}
}
\usage{
exploredata_ex
}
\description{
Demo data for exploredata
}
\keyword{datasets}
