% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module.R
\name{exploredata_ui}
\alias{exploredata_ui}
\alias{exploredata_server}
\title{Shiny module for explore data}
\usage{
exploredata_ui(id, label = "Variable :")

exploredata_server(id, data)
}
\arguments{
\item{id}{: module id}

\item{label}{: \code{character} name of column input}

\item{data}{: \code{data.frame} to explore}
}
\value{
\code{NULL}
}
\description{
Shiny module for explore data
}
\examples{

\dontrun{

data("exploredata_ex")

ui <- fluidPage(
  exploredata_ui("id")
)

server <- function(input, output, session) {
  exploredata_server("id",
    reactive(exploredata_ex)
  )
}

shinyApp(ui, server)

}

}
