% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module.R
\name{run_explore_data_app}
\alias{run_explore_data_app}
\title{Run explore data shiny app}
\usage{
run_explore_data_app(data)
}
\arguments{
\item{data}{: a \code{data.frame}}
}
\description{
Run explore data shiny app
}
\examples{

\dontrun{

run_explore_data_app(iris)

}

}
